refactor: modifica importación de carpeta utils y enums
