+++ src/main/java/com/tienda/facil/core/model/Pedido.java
@@ -1,7 +1,7 @@
 package com.tienda.facil.core.model;
 
-import com.tienda.facil.core.utils.enums.EstadoPedido;
-import com.tienda.facil.core.utils.enums.MetodoPago;
+import com.tienda.facil.core.util.enums.EstadoPedido;
+import com.tienda.facil.core.util.enums.MetodoPago;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -34,7 +34,7 @@ public class PedidoModel {
      */
     @ManyToOne
     @JoinColumn(name = "cliente_id", referencedColumnName = "id")
-    private ClienteModel cliente;
+    private Cliente cliente;
 
     /**
      * Relación muchos a uno con Prioridad.
