--- src/main/java/com/tienda/facil/core/utils/Constantes.java
@@ -1,6 +1,0 @@
-package com.tienda.facil.core.utils;
-
-public class Constantes {
-    public static final String PEDIDO_NO_ENCONTRADO = "Pedido no encontrado";
-    public static final String PEDIDOS_OBTENIDOS = "Pedidos obtenidos exitosamente";
-}
